[OOTDay-refactoring] 파일 정리 및 Styling DTO 추가
