+++ src/main/java/TOTs/OOTDay/styling/stylingoption/repository/MoodRepository.java
@@ -1,6 +1,6 @@
-package TOTs.OOTDay.util.stylingoption.repository;
+package TOTs.OOTDay.styling.stylingoption.repository;
 
-import TOTs.OOTDay.util.stylingoption.domain.Mood;
+import TOTs.OOTDay.styling.stylingoption.domain.Mood;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MoodRepository extends JpaRepository<Mood, Long> {
