+++ src/main/java/TOTs/OOTDay/styling/stylingoption/repository/PlaceRepository.java
@@ -1,6 +1,6 @@
-package TOTs.OOTDay.util.stylingoption.repository;
+package TOTs.OOTDay.styling.stylingoption.repository;
 
-import TOTs.OOTDay.util.stylingoption.domain.Place;
+import TOTs.OOTDay.styling.stylingoption.domain.Place;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PlaceRepository extends JpaRepository<Place, Long> {
