+++ src/main/java/TOTs/OOTDay/wardrobe/ClothCategory.java
@@ -1,4 +1,4 @@
- package TOTs.OOTDay.domain;
+ package TOTs.OOTDay.wardrobe;
 
 import com.fasterxml.jackson.annotation.JsonValue;
